requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
chromadb
sentence-transformers
faiss-cpu
numpy

